Kevin_update_test1213

+++ Mywebapp/src/main/java/com/mycompany/user/UserRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface UserRepository extends CrudRepository<User, Integer> {
   public Long countById(Integer id);
-  public int countByAccount(String account);
+  public Long countByAccount(String account);
   public User findByAccount(String account);
 }
